requirements.txt
openai
SpeechRecognition
pyttsx3
pyaudio
python-dotenv
